Split junction relation resolver tests
